Extract panel markup and shared constants out of App.ts
First step of the agreed v0.6 App.ts decomposition (8,149 lines), taken
in risk order: this commit moves only code with no runtime behavior.

appMarkup.ts takes the createAppMarkup HTML builders, getAppElements,
getElement, and the AppElements type. appConstants.ts takes the shared
configuration constants (defaults, fallbacks, version) and the tool-card
catalog with its types, which both App.ts and the markup reference.

Every moved block was verified byte-identical to its original after the
move; the only changes are export keywords and import statements. App.ts
drops from 8,149 to 6,533 lines.
